Add mappable data support to GeoPointMapFragment
